chore: create posthog-edge package
