+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -20,6 +20,7 @@ _Please describe._
 - [ ] posthog-web
 - [ ] posthog-node
 - [ ] posthog-ai
+- [ ] posthog-edge
 - [ ] posthog-react-native
 
 ## Additional context
