+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -18,6 +18,7 @@ _Please describe._
 - [ ] posthog-web
 - [ ] posthog-node
 - [ ] posthog-ai
+- [ ] posthog-edge
 - [ ] posthog-react-native
 
 ## Additional context
