+++ .github/workflows/release.yml
@@ -20,6 +20,8 @@ jobs:
             npm_token_secret: NPM_TOKEN # TODO: Change to standard token once able
           - name: posthog-ai
             npm_token_secret: NPM_TOKEN # TODO: Change to standard token once able
+          - name: posthog-edge
+            npm_token_secret: NPM_TOKEN # TODO: Change to standard token once able
 
     steps:
       - name: Checkout the repository
