+++ CHANGELOG.md
@@ -6,5 +6,6 @@ Find the changelogs for the respective libraries in their folders:
 2. [posthog-react-native](https://github.com/PostHog/posthog-js-lite/blob/main/posthog-react-native/CHANGELOG.md)
 3. [posthog-web](https://github.com/PostHog/posthog-js-lite/blob/main/posthog-web/CHANGELOG.md)
 4. [posthog-ai](https://github.com/PostHog/posthog-js-lite/blob/main/posthog-ai/CHANGELOG.md)
+5. [posthog-edge](https://github.com/PostHog/posthog-js-lite/blob/main/posthog-edge/CHANGELOG.md)
 
 For Session Replay, See [posthog-react-native-session-replay](https://github.com/PostHog/posthog-react-native-session-replay/blob/main/CHANGELOG.md)
