+++ CONTRIBUTING.md
@@ -9,6 +9,7 @@ This repository is broken into different packages
 - **/posthog-react-native** > React Native specific code
 - **/posthog-web** > Web (DOM) specific code
 - **/posthog-ai** > Node.js SDK for LLM Observability
+- **/posthog-edge** > Edge compatible version of the Node.js SDK
 
 For Session Replay, See [posthog-react-native-session-replay](https://github.com/PostHog/posthog-react-native-session-replay/CHANGELOG.md)
