+++ extensions/error-tracking/autocapture.ts
@@ -1,7 +1,7 @@
 // Portions of this file are derived from getsentry/sentry-javascript by Software, Inc. dba Sentry
 // Licensed under the MIT License
 
-import { EventHint } from 'posthog-node/src/extensions/error-tracking/types'
+import { EventHint } from 'posthog-node/extensions/error-tracking/types'
 
 type ErrorHandler = { _posthogErrorHandler: boolean } & ((error: Error) => void)
