+++ extensions/error-tracking/reduceable-cache.ts
