+++ extensions/error-tracking/stack-trace.ts
