+++ extensions/error-tracking/type-checking.ts
