+++ extensions/error-tracking/types.ts
