+++ posthog-edge/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../tsconfig.json",
+    "compilerOptions": {
+        "types": ["node"],
+        "typeRoots": ["./node_modules/@types", "../node_modules/@types"]
+    }
+}
