+++ posthog-node/src/types.ts
@@ -1,70 +1,5 @@
 import { FeatureFlagValue, JsonType } from '../../posthog-core/src'
 
-export interface IdentifyMessage {
-  distinctId: string
-  properties?: Record<string | number, any>
-  disableGeoip?: boolean
-}
-
-export interface EventMessage extends IdentifyMessage {
-  event: string
-  groups?: Record<string, string | number> // Mapping of group type to group id
-  sendFeatureFlags?: boolean
-  timestamp?: Date
-  uuid?: string
-}
-
-export interface GroupIdentifyMessage {
-  groupType: string
-  groupKey: string // Unique identifier for the group
-  properties?: Record<string | number, any>
-  distinctId?: string // optional distinctId to associate message with a person
-  disableGeoip?: boolean
-}
-
-export type PropertyGroup = {
-  type: 'AND' | 'OR'
-  values: PropertyGroup[] | FlagProperty[]
-}
-
-export type FlagProperty = {
-  key: string
-  type?: string
-  value: string | number | (string | number)[]
-  operator?: string
-  negation?: boolean
-}
-
-export type FeatureFlagCondition = {
-  properties: FlagProperty[]
-  rollout_percentage?: number
-  variant?: string
-}
-
-export type PostHogFeatureFlag = {
-  id: number
-  name: string
-  key: string
-  filters?: {
-    aggregation_group_type_index?: number
-    groups?: FeatureFlagCondition[]
-    multivariate?: {
-      variants: {
-        key: string
-        rollout_percentage: number
-      }[]
-    }
-    payloads?: Record<string, string>
-  }
-  deleted: boolean
-  active: boolean
-  /** @deprecated This field will be removed in a future version. **/
-  is_simple_flag: boolean
-  rollout_percentage: null | number
-  ensure_experience_continuity: boolean
-  experiment_set: number[]
-}
-
 export type PostHogNodeV1 = {
   /**
    * @description Capture allows you to capture anything a user does within your system,
